+++ empire/test/test_listener_api.py
@@ -100,14 +100,17 @@ def test_create_listener_validation_fails_required_field(client, admin_auth_head
 
 
 def test_create_listener_custom_validation_fails(client, admin_auth_header):
-    base_listener = get_base_listener()
+    base_listener = get_base_malleable_listener()
     base_listener["name"] = "temp123"
-    base_listener["options"]["Host"] = "https://securedomain.com"
+    base_listener["options"]["Profile"] = "nonexistent.profile"
     response = client.post(
         "/api/v2/listeners/", headers=admin_auth_header, json=base_listener
     )
     assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.json()["detail"] == "[!] HTTPS selected but no CertPath specified."
+    assert (
+        response.json()["detail"]
+        == "[!] Malleable profile not found: nonexistent.profile"
+    )
 
 
 def test_create_listener_template_not_found(client, admin_auth_header):
@@ -194,6 +197,28 @@ def test_create_listener_normalization_sets_host_port_as_bind_port(
     assert response.status_code == status.HTTP_201_CREATED
     assert response.json()["options"]["Host"] == "http://localhost:443"
     assert response.json()["options"]["Port"] == "443"
+
+    client.delete(
+        f"/api/v2/listeners/{response.json()['id']}", headers=admin_auth_header
+    )
+
+
+def test_create_listener_with_https_host_no_cert_path(client, admin_auth_header):
+    # A listener should be able to have a host with https even if the cert path is blank.
+    # because the listener might be behind a reverse proxy that handles TLS.
+    base_listener = get_base_listener()
+    base_listener["name"] = "temp123"
+    base_listener["options"]["Host"] = "https://localhost:443"
+    base_listener["options"]["Port"] = "80"
+    base_listener["options"]["CertPath"] = ""
+    response = client.post(
+        "/api/v2/listeners/", headers=admin_auth_header, json=base_listener
+    )
+
+    assert response.status_code == status.HTTP_201_CREATED
+    assert response.json()["options"]["Host"] == "https://localhost:443"
+    assert response.json()["options"]["Port"] == "80"
+    assert response.json()["options"]["CertPath"] == ""
 
     client.delete(
         f"/api/v2/listeners/{response.json()['id']}", headers=admin_auth_header
@@ -406,30 +431,51 @@ def test_update_listener_reverts_if_validation_fails(
 
 
 def test_update_listener_reverts_if_custom_validation_fails(
-    client, admin_auth_header, listener
+    client, admin_auth_header, listener_malleable
 ):
+    listener_malleable["enabled"] = False
+    response = client.put(
+        f"/api/v2/listeners/{listener_malleable['id']}",
+        headers=admin_auth_header,
+        json=listener_malleable,
+    )
+    assert response.status_code == status.HTTP_200_OK
+    assert response.json()["enabled"] is False
+
     response = client.get(
-        f"/api/v2/listeners/{listener['id']}",
+        f"/api/v2/listeners/{listener_malleable['id']}",
         headers=admin_auth_header,
     )
     assert response.json()["enabled"] is False
 
-    listener = response.json()
-    listener["options"]["Host"] = "https://securesite.com"
-    listener["options"]["BindIP"] = "1.1.1.1"
+    listener_malleable = response.json()
+    listener_malleable["options"]["Profile"] = "nonexistent.profile"
     response = client.put(
-        f"/api/v2/listeners/{listener['id']}",
+        f"/api/v2/listeners/{listener_malleable['id']}",
         headers=admin_auth_header,
-        json=listener,
+        json=listener_malleable,
     )
     assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.json()["detail"] == "[!] HTTPS selected but no CertPath specified."
+    assert (
+        response.json()["detail"]
+        == "[!] Malleable profile not found: nonexistent.profile"
+    )
 
     response = client.get(
-        f"/api/v2/listeners/{listener['id']}",
+        f"/api/v2/listeners/{listener_malleable['id']}",
         headers=admin_auth_header,
     )
-    assert response.json()["options"]["BindIP"] == "0.0.0.0"
+    assert response.json()["options"]["Profile"] == "amazon.profile"
+
+    listener_malleable["options"]["Profile"] = "amazon.profile"
+    listener_malleable["enabled"] = True
+    response = client.put(
+        f"/api/v2/listeners/{listener_malleable['id']}",
+        headers=admin_auth_header,
+        json=listener_malleable,
+    )
+
+    assert response.status_code == status.HTTP_200_OK
 
 
 def test_update_listener_allows_and_enables_while_disabled(